requirements.txt
camelot-py==0.10.1
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cryptography==36.0.0
decorator==5.1.0
Elixir==0.7.1
et-xmlfile==1.1.0
ghostscript==0.7
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
opencv-python==4.5.4.60
openpyxl==3.0.9
pandas==1.3.4
pbr==5.8.0
pdfminer.six==20211012
pdftopng==0.2.3
pycparser==2.21
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.9
Tempita==0.5.2
xlrd==0.7.1
xlwt==0.7.2
